pyproject.toml
[project]
name = "drive-backup-threading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.178.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]
